requirements.txt
flask
flask-restful
flask_cors
pdf2image
torch
torchvision
opencv-python
scikit-image
scipy
gitpython
matplotlib
numpy
Pillow
psutil  # system resources
PyYAML
requests
thop  # FLOPs computation
tqdm
faiss-cpu
six
tensorflow
pandas
vietocr
paddleocr
paddlepaddle